UI updates , employee form fixed

+++ app/master/routes.py
@@ -180,15 +180,20 @@ def delete_location():
         print(payload)
         check_data = Location.query.filter_by(id=payload['id'])
         if check_data.first():
-            try:
-                print(check_data.first().name)
-                check_data.delete()
-                db.session.commit()
-                return jsonify({'success': 'Data deleted'})
-            except Exception as e:
-                db.session.rollback()
-                db.session.close()
-                return jsonify({'message': 'Something unexpected happened. Check logs', 'log': str(e)})
+            if len(check_data.first().company_location) is int(0):
+
+                try:
+                    print(check_data.first().name)
+                    check_data.delete()
+                    db.session.commit()
+                    return jsonify({'success': 'Data deleted'})
+                except Exception as e:
+                    db.session.rollback()
+                    db.session.close()
+                    return jsonify({'message': 'Something unexpected happened. Check logs', 'log': str(e)})
+            else:
+                return jsonify({'message': 'Cannot delete data. Being used by other master.'})
+
         else:
             return jsonify({'message': 'Data does not exist.'})
